Move XPointParseError from library to common exceptions
The XPoint value object lives in domain/common but raised an exception
defined in domain/library, giving every module that touches
common.value_objects a transitive dependency on library.
