+++ backend/src/application/reading/services/highlight_grouping_service.py
