--- backend/src/domain/library/services/__init__.py
@@ -1,1 +1,0 @@
-"""Library domain services."""
